Merge penguin:v2.5/linux into home.transmeta.com:/home/torvalds/v2.5/linux
2002/06/13 10:12:12-07:00 axboe
[PATCH] generic tag update

The following comes from James Bottemley, as he updated the SCSI layer
to use the generic tagging.

Basically just put blk_queue_tag_request() into ll_rw_blk.c and name it
something more sane, blk_queue_find_tag(). In addition, remove REQ_CMD
requirement in blk_queue_start_tag() -- this is actually IDE specific
(we can only tag read/write requests there...), SCSI actually requires
everything to be tagged once enabled. It's replaced with a safety check
for an already tagged request.

2002/06/13 10:11:59-07:00 axboe
[PATCH] final plug stuff

This is a merge of the two things that are needed for the plugging as it
stands in 2.5.21. The first is fixing the locking to be both clearer and
safe (Andrew repeatedly broke the old version). The second is a few
changes that allow make_request_fn drivers to utilize plugging. This is
needed for umem and raid, for instance, that have their private
plugging.

2002/06/13 10:11:39-07:00 torvalds
Cleanup, use new bitmap declarator instead of doing it by hand.

BKrev: 3d08d288M8ANF2TMDTMjqXU6I5ZKIg