[PATCH] 2.5.20 x86 iobitmap cleanup
This makes the IO bitmap allocations a separately allocated structure,
shrinking the default task size.

We alloc it in sys_ioperm() and copy_thread() and frees in
exit_thread().  It also gets rid of the IO_BITMAP_SIZE+1 crap, as only
the tss actually needs the tail long, and we weren't copying it into the
bitmap anyways.

BKrev: 3d08d800CI58W0ptGjmfAm5P9lAbNA