[PATCH] mark_buffer_dirty_inode() speedup
buffer_insert_list() is showing up on Anton's graphs.  It'll be via
ext2's mark_buffer_dirty_inode() against indirect blocks.  If the
buffer is already on an inode queue, we know that it is on the correct
inode's queue so we don't need to re-add it.

BKrev: 3d0f6b74MSjvLd137s6RenyLNQa8vw