[PATCH] Make ITERATE_MDDEV work on non-SMP
For an SMP kernel, spin_lock() et.al. are functions.
For a UP kernel, they are statements that must be terminated
by a ';'.  This is not quite the same thing, and hence spin_lock()
cannot be using inside a parenthesised expession.

This patch changes ITERATE_MDDEV to use gcc's "statement expressions"
instead which has the benefit of making the conditionals
more readable.

BKrev: 3d122bb4OXGol-AKJNlyDyD10-iOrA