Merge kroah.com:/home/linux/linux/BK/bleeding-2.5 into kroah.com:/home/linux/linux/BK/gregkh-2.5
2002/07/04 23:46:47-07:00 petkan
[PATCH] pegasus & rtl8150

I chose a little bit more restrictive license for my drivers.
Rx skb pool introduced in pegasus driver and the pool locking in rtl8150
is refined.

2002/07/04 18:21:24-07:00 mdharm-usb
[PATCH] usb-storage: remove timer

This removes the timer usage in usb-storage.  This cleans up quite a bit
of the state machine and eliminates quite a few potential races.

Initialization commands and other non-data-path mechanisms use the USB core
timeout mechanism.  Anything in the data path uses the SCSI mid-layer
mechanism.

2002/07/04 18:21:18-07:00 mdharm-usb
[PATCH] usb-storage: Code consolidation of error paths

This patch consolidates quite of bit of code in the control thread to
place all the cleanup/error handling into one place.

2002/07/04 18:21:08-07:00 mdharm-usb
[PATCH] usb-storage: code cleanup, small fixes

This patch consolidates quite a bit of code for allocation/deallocation of
URBs, and removes a kmalloc() from a command path.

2002/07/04 11:04:13-07:00 greg
USB:  removed unused Config.help entries from the host controller file.

2002/07/04 11:03:26-07:00 greg
USB:  picked a uhci driver to go forward with.

Removed usb-uhci-hcd.o from the list of UHCI drivers.
This allowed the logic to be cleaned up.
Removed CONFIG_EXPERIMENTAL dependancy, as it's no longer needed.

2002/06/21 15:12:01-07:00 david-b
[PATCH] Re: [linux-usb-devel] unending timeouts (patch for 2.5.22 oops)

Ah, so both of the "hcd-ized" UHCI drivers have a common bug:
they've got logic to look at the USB_ASYNC_UNLINK flag and
block unless it's clear ... but the hcd framework is already
handling the synchronous behavior, so that's wrong.

Try to repeat that with the patch I've attached, which rips
out that duplicated code ... and so should at least get rid of
that oops, even if it doesn't entirely fix the timeout issue.

(Or: try with either the OHCI driver, or with the EHCI driver
through a USB 2.0 hub, if you have appropriate hardware.)

- Dave

p.s. Disclaimer about this patch:  all it does is rip out
      code and make it compile without warnings, but I've
      not tested it otherwise.  There's a possiblity it'll
      uncover latent issues on the other code path, but then
      that's exactly why we only want one unlink code path
      inside the HCDs!  So Greg, please merge anyway ...

2002/06/21 15:11:45-07:00 david-b
[PATCH] ohci-hcd cardbus unplug

This is the 2.5 version of that 2.4 patch I sent recently,
which makes the OHCI driver behave usably on at least
some cardbus systems when the card is just ejected without
a clean shutdown.

BKrev: 3d25ff2b1yTxvzE0SSIie3oZS2gzhw