[PATCH] Fix note sections in ELF core dumps
Edition 4.1 of the System V Application Binary Interface says that
"The first namesz bytes in name contains a null-terminated
representation of the entry's owner or originator".  This implies that
the terminating null is included in namesz, which is corroborated by
the example that follows the description.  However, this is not what
the Linux kernel does when it writes its notes into an ELF core dump.
The attached patch fixes this.

BKrev: 3d262b41zqvXV7QMkcwp940SzOp6Qg