Merge http://linuxusb.bkbits.net/linus-2.5 into home.transmeta.com:/home/torvalds/v2.5/linux
2002/07/15 17:49:09-07:00 greg
USB: fix flag name in ohci driver due to previous patch.

2002/07/15 17:47:21-07:00 david-b
[PATCH] ohci misc

This patch  includes the innocuous bits from a larger one that
I'm still working on (mostly unlink fixes):

- updates comments
- flags TDs that were seen in the donelist
- removes some bogus whitespace (at EOL etc) and tabs
- checks for an enumeration issue that might cause trouble
- delays IRQs a bit more aggressively
- shortens TD submit paths a smidgeon (smaller ".o")
- updates some of the debug output
- sanitizes usb_make_path() output on the SA-1111

2002/07/15 17:37:47-07:00 vojtech
[PATCH] A cleanup of Paul's 2.5 hiddev update.

Get rid of #ifdefs in hid-core again. (For you, Greg.)
Move the uref generation code from hid-core to hiddev to make things cleaner.

2002/07/15 17:19:17-07:00 mdharm-usb
[PATCH] usb-storage: catch bad commands

The purpose of this patch is to trap all commands which have a bogus
request_bufflen.  Much logic is devoted to calculating the proper length of
the transfer, but according to discussions I've had on linux-scsi, this is
really a bug in whatever is originating the bad command.

Hopefully, after people use this patch for a while, we'll eliminate all the
offending sources and can remove quite a bit of logic from the driver.

2002/07/15 17:18:59-07:00 vojtech
[PATCH] Reintroduce proper returning of -EFAULT to hiddev.c

One of my earlier patches backed out a change that fixed all
copy_to_user calls to properly return -EFAULT if needed. This puts the
change back in.

2002/07/15 17:18:42-07:00 vojtech
[PATCH] Updates for hiddev by Paul Stewart

I've merged a patch Paul Stewart sent me some time ago, which should make life
easier for the guys writing UPS daemons.

BKrev: 3d337469pf3ErqDELZNV8bIZPfgi3A