[PATCH] RPC over UDP congestion control updates [2/8]
Implement a count of the number of timeouts that have occured since
we last recorded a successful reply from the server.

For the moment this information is merely used in order to improve the
estimate of whether or not the server is down. It will be used in
patch 3/8 in order to improve the timeout backoff algorithm.

BKrev: 3d34a3f7vmYDYwVEvP2ahqA4rze4Pw