[PATCH] RPC over UDP congestion control updates [3/8]
Improve the response to timeouts. As requests time out, we delay
timing out the remaining requests (in fact we follow exponential
backoff). This is done because we assume either that the round trip
time has been underestimated, or that the network/server is congested,
and we need to back off the resending of new requests.

BKrev: 3d34a404Q-WRM6D5qhGLU0I-WCRX-Q