[PATCH] MD - Don't maintain disc status in superblock.
Don't maintain disc status in superblock.

The state is now in rdev so we don't maintain it
in superblock any more.
We also nolonger test content of superblock for
disk status
mddev->spare is now an rdev and not a superblock fragment.

BKrev: 3d3839e86_Z8NN3UPJ-UTEmWPT4I7A