[PATCH] MD - Remove dependance on superblock
Remove dependance on superblock

All the remaining field of interest in the superblock
get duplicated in the mddev struture and this is treated as
authoritative.  The superblock gets completely generated at
write time, and all useful information extracted at read time.

This means that we can slot in different superblock formats
without affecting the bulk of the code.

BKrev: 3d383a2fvTk82Wm63PXFKPofjSEJrQ