[PATCH] avoid allocating pte_chains for unshared pages
Patch from David McCracken.  It is an optimisation to the rmap
pte_chains.

In the common case where a page is mapped by only a single pte, we
don't need to allocate a pte_chain structure.  Just make the page's
pte_chain pointer point straight at that pte and flag this with
PG_direct.

BKrev: 3d3855beGsdR-7MVgWd5vjBB9zqxOg