[PATCH] MD - Fix two bugs that would cause sync_sbs to Oops
Sync_sbs tries to access the ->sb for the first rdev of an mddev.
This can oops as the wrong arg is given to list_entry, and also
if a define was faound to be failed, as failed devices have their ->sb
removed.  But that removal isn't necessary, so now an rdev will always
have an ->sb.

BKrev: 3d3ee68eweEHMxRnMu_7VutBswyCoA