Merge bk://bk.arm.linux.org.uk:14691 into home.transmeta.com:/home/torvalds/v2.5/linux
2002/07/23 22:31:48+01:00 rmk
[SERIAL] Fix sa1100 serial driver stop function parameters.

2002/07/23 08:15:28+01:00 rmk
[SERIAL] Fix deadlock in __uart_start introduced in previous cset
Thanks to Zwane Mwaikambo for finding this.

2002/07/23 00:45:58+01:00 rmk
[SERIAL] Fix another SMP deadlock with modem status signal changes
The original fix sent to Ingo for stop_tx didn't take account that
the start_tx and stop_tx methods can be called from the device
specific code under the port spinlock.  Consequently, we move the
spinlock to the callers of these methods.  Documentation updated
to reflect the change.

2002/07/22 10:41:11+01:00 rmk
[SERIAL] Fix up various filenames, etc, from Ingo's merge of serial.

2002/07/22 10:32:03+01:00 rmk
[SERIAL] Rename files to remove serial_ prefix.

BKrev: 3d3ee819TACsQTxikgVurkkadskOuQ